+++ src/eg3dproj.h
@@ -4,6 +4,6 @@
 #include "inttype.h"
 
 uint32 scaleCoordToLod(int16 level, uint32 coord);
-void projectObjects(int16 heading, int16 rangeGate, long worldX, long worldY, long worldZ);
+void projectObjects(int16 heading, int16 rangeGate, int32 worldX, int32 worldY, int32 worldZ);
 
 #endif /* F15_SE2_EG3DPROJ */
