+++ src/egmath.h
@@ -4,7 +4,7 @@
 /* public interface of egmath.c */
 
 void load15Flt3d3();
-void drawWorldObject(int16 shapeId, long worldX, long worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 scaleShift);
+void drawWorldObject(int16 shapeId, int32 worldX, int32 worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 scaleShift);
 void drawTargetView(int16 shapeId, int16 worldX, int16 worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 mode, int16 shift);
 int16 shapeDataOffset(int16 shapeId);
 int16 clampRange(int16 value, int16 minVal, int16 maxVal);
